Making sure timers do not leak
Timer does not retain self anymore.
We now have an invalidate method that clears the source event handler
and then resumes the source if suspeneded and then nilifies the source.
Earlier we when we called clear, it was suspending the source and then
we tried to resume it in dealloc. Now, we skip the suspend/resume pair
for the common case of invalidate followed by dealloc.